requirements.txt
# Core Framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0

# LangChain and LLM
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
groq>=0.4.0

# Embeddings and Vector Store
sentence-transformers>=2.3.0
chromadb>=0.4.22

# Additional Dependencies
numpy>=1.24.0
requests>=2.31.0
aiofiles>=23.2.0
python-multipart>=0.0.6

# HTTP and CORS
httpx>=0.26.0
